requirements.txt
numpy~=1.23.0
Pillow~=9.4.0
tqdm~=4.66.1
matplotlib~=3.7.1
scikit-learn~=1.2.2
pandas~=1.5.3
opencv-python~=4.6.0.66